requirements.txt
pygame==2.6.0
PyYAML